+++ packages/@expo/cli/src/run/ios/runIosAsync.ts
@@ -13,6 +13,7 @@ import { profile } from '../../utils/profile';
 import { getSchemesForIosAsync } from '../../utils/scheme';
 import { ensureNativeProjectAsync } from '../ensureNativeProject';
 import { logProjectLogsLocation } from '../hints';
+import { uploadRemoteBuildCache, resolveRemoteBuildCache } from '../remoteBuildCache';
 import { startBundlerAsync } from '../startBundler';
 import * as XcodeBuild from './XcodeBuild';
 import { Options } from './XcodeBuild.types';
@@ -22,7 +23,6 @@ import { getValidBinaryPathAsync } from './validateExternalBinary';
 import { exportEagerAsync } from '../../export/embed/exportEager';
 import { getContainerPathAsync, simctlAsync } from '../../start/platforms/ios/simctl';
 import { CommandError } from '../../utils/errors';
-import { resolveRemoteBuildCache } from '../remoteBuildCache';
 
 const debug = require('debug')('expo:run:ios');
 
@@ -102,6 +102,7 @@ export async function runIosAsync(projectRoot: string, options: Options) {
   }
 
   let binaryPath: string;
+  let shouldUpdateBuildCache = false;
   if (options.binary) {
     binaryPath = await getValidBinaryPathAsync(options.binary, props);
     Log.log('Using custom binary path:', binaryPath);
@@ -126,6 +127,7 @@ export async function runIosAsync(projectRoot: string, options: Options) {
     // Find the path to the built app binary, this will be used to install the binary
     // on a device.
     binaryPath = await profile(XcodeBuild.getAppBinaryPath)(buildOutput);
+    shouldUpdateBuildCache = true;
   }
   debug('Binary path:', binaryPath);
 
@@ -178,6 +180,13 @@ export async function runIosAsync(projectRoot: string, options: Options) {
     logProjectLogsLocation();
   } else {
     await manager.stopAsync();
+  }
+
+  if (shouldUpdateBuildCache && projectConfig.exp.experiments?.remoteBuildCache) {
+    await uploadRemoteBuildCache(projectRoot, {
+      platform: 'ios',
+      provider: projectConfig.exp.experiments?.remoteBuildCache.provider,
+    });
   }
 }